refactor(#191): split check_template_bundles into fetch/validate/config modules
The hook lived in one ~536-line module mixing four responsibilities. Split the
internals into focused private modules, keeping check_template_bundles.py as the
thin CLI/orchestration layer that wires them together and re-exports the helpers
so `rhiza_hooks.check_template_bundles` stays the single public import surface:

- _bundles_fetch.py    BundlesDoc, fetch constants, local/remote loading + retry
- _bundles_validate.py structural validators + validate_template_bundles
- _bundles_config.py   .rhiza/template.yml readers
- check_template_bundles.py  main(), argparse, orchestration, re-exports

No behavior change. The `check-template-bundles` entry point is unchanged.
Test monkeypatch targets for urlopen/urlparse/time (looked up inside the fetch
function) move to rhiza_hooks._bundles_fetch; _fetch_remote_bundles patches stay
valid via the imported name in the CLI module. 299 tests pass, 100% line+branch
coverage, mypy --strict and ty green.

Closes #191
