+++ src/rhiza_hooks/_bundles_validate.py
@@ -1,0 +1,186 @@
+#!/usr/bin/env python3
+"""Structural validation of a template-bundles document.
+
+These functions operate on an already-loaded mapping (see
+:mod:`rhiza_hooks._bundles_fetch` for how the document is obtained) and report
+problems as lists of human-readable error strings. They never perform I/O,
+except :func:`validate_template_bundles`, which is a thin convenience wrapper
+that loads a local file and then validates it.
+"""
+
+from __future__ import annotations
+
+from collections.abc import Callable
+from pathlib import Path
+from typing import Any
+
+from rhiza_hooks._bundles_fetch import _load_yaml_file
+
+
+def _validate_top_level_fields(data: dict[Any, Any]) -> list[str]:
+    """Validate required top-level fields."""
+    errors = []
+    required_fields = {"version", "bundles"}
+    for field in required_fields:
+        if field not in data:
+            errors.append(f"Missing required field: {field}")
+    return errors
+
+
+def _validate_bundle_structure(
+    bundle_name: str,
+    bundle_config: Any,
+    bundle_names: set[str],
+) -> list[str]:
+    """Validate a single bundle's structure and dependencies."""
+    errors = []
+
+    if not isinstance(bundle_config, dict):
+        errors.append(f"Bundle '{bundle_name}' must be a dictionary")
+        return errors
+
+    # Check required fields
+    if "description" not in bundle_config:
+        errors.append(f"Bundle '{bundle_name}' missing 'description'")
+
+    if "files" not in bundle_config:
+        errors.append(f"Bundle '{bundle_name}' missing 'files'")
+    elif not isinstance(bundle_config["files"], list):
+        errors.append(f"Bundle '{bundle_name}' 'files' must be a list")
+
+    # Validate dependencies
+    if "requires" in bundle_config:
+        if not isinstance(bundle_config["requires"], list):
+            errors.append(f"Bundle '{bundle_name}' 'requires' must be a list")
+        else:
+            for dep in bundle_config["requires"]:
+                if dep not in bundle_names:
+                    errors.append(f"Bundle '{bundle_name}' requires non-existent bundle '{dep}'")
+
+    if "recommends" in bundle_config:
+        if not isinstance(bundle_config["recommends"], list):
+            errors.append(f"Bundle '{bundle_name}' 'recommends' must be a list")
+        else:
+            for dep in bundle_config["recommends"]:
+                if dep not in bundle_names:
+                    errors.append(f"Bundle '{bundle_name}' recommends non-existent bundle '{dep}'")
+
+    return errors
+
+
+def _validate_examples(examples: Any, bundle_names: set[str]) -> list[str]:
+    """Validate examples section."""
+    errors = []
+
+    if not isinstance(examples, dict):
+        errors.append("'examples' must be a dictionary")
+        return errors
+
+    for example_name, example_config in examples.items():
+        if "templates" in example_config:
+            if not isinstance(example_config["templates"], list):
+                errors.append(f"Example '{example_name}' 'templates' must be a list")
+            else:
+                for template in example_config["templates"]:
+                    # core is auto-included, we don't validate it
+                    if template != "core" and template not in bundle_names:
+                        errors.append(f"Example '{example_name}' references non-existent bundle '{template}'")
+
+    return errors
+
+
+def _validate_metadata(metadata: dict[Any, Any], bundles: dict[Any, Any]) -> list[str]:
+    """Validate metadata section."""
+    errors = []
+
+    if "total_bundles" in metadata:
+        expected_count = len(bundles)
+        actual_count = metadata["total_bundles"]
+        if actual_count != expected_count:
+            errors.append(
+                f"Metadata 'total_bundles' ({actual_count}) doesn't match actual bundle count ({expected_count})"
+            )
+
+    return errors
+
+
+def _validate_selected_bundles(
+    templates: set[str],
+    bundles: dict[Any, Any],
+    missing_message: Callable[[str], str],
+) -> list[str]:
+    """Validate that each requested template exists in ``bundles`` and is well-formed.
+
+    Shared by the local-file and remote-fetch paths; they differ only in the
+    "not found" wording, supplied via ``missing_message``.
+
+    Args:
+        templates: Template names requested in the rhiza config.
+        bundles: The ``bundles`` mapping from a template-bundles document.
+        missing_message: Builds the error string for a template absent from ``bundles``.
+
+    Returns:
+        List of error messages (empty if every requested template is valid).
+    """
+    errors: list[str] = []
+    bundle_names: set[str] = set(bundles.keys())
+
+    for template in templates:
+        if template not in bundle_names:
+            errors.append(missing_message(template))
+
+    for template in templates:
+        if template in bundles:
+            errors.extend(_validate_bundle_structure(template, bundles[template], bundle_names))
+
+    return errors
+
+
+def validate_template_bundles(bundles_path: Path, templates_to_check: set[str] | None = None) -> tuple[bool, list[str]]:
+    """Validate template bundles configuration.
+
+    Args:
+        bundles_path: Path to template-bundles.yml
+        templates_to_check: Optional set of template names to validate. If None, validate all.
+
+    Returns:
+        Tuple of (success, error_messages)
+    """
+    # Load YAML file
+    loaded = _load_yaml_file(bundles_path)
+    if loaded.data is None:
+        return False, loaded.errors
+    # data is narrowed to dict[Any, Any] by the `is None` guard above.
+    data = loaded.data
+
+    # Validate top-level fields
+    errors = _validate_top_level_fields(data)
+    if errors:
+        return False, errors
+
+    # Validate bundles section
+    bundles = data.get("bundles", {})
+    if not isinstance(bundles, dict):
+        return False, ["'bundles' must be a dictionary"]
+
+    bundle_names: set[str] = {str(k) for k in bundles}
+
+    if templates_to_check is not None:
+        # Validate only the requested subset (existence + structure).
+        errors.extend(
+            _validate_selected_bundles(
+                templates_to_check,
+                bundles,
+                lambda t: f"Template '{t}' specified in .rhiza/template.yml not found in bundles",
+            )
+        )
+    else:
+        # Validate every declared bundle, plus the examples and metadata sections.
+        for bundle_name in bundle_names:
+            errors.extend(_validate_bundle_structure(bundle_name, bundles[bundle_name], bundle_names))
+        if "examples" in data:
+            errors.extend(_validate_examples(data["examples"], bundle_names))
+        if "metadata" in data:
+            errors.extend(_validate_metadata(data["metadata"], bundles))
+
+    return len(errors) == 0, errors
